Fix Dual XDMA Test Environment
